+++ src/statics.rs
@@ -1,3 +1,5 @@
+use std::num::{NonZeroU8, NonZeroUsize};
+
 pub struct VersionInfo<'a> {
     pub commit: &'a str,
     pub semver: &'a str,
@@ -8,6 +10,6 @@ pub const VERSION_INFO: VersionInfo = VersionInfo {
     semver: env!("VERGEN_GIT_SEMVER_LIGHTWEIGHT"),
 };
 
-pub static DEFAULT_SEPARATOR: &str = "";
-pub static DEFAULT_WORDS_PER_NAME: u8 = 2;
-pub static DEFAULT_NUMBER_OF_NAMES: usize = 6;
+pub const DEFAULT_SEPARATOR: &str = "";
+pub const DEFAULT_WORDS_PER_NAME: NonZeroU8 = unsafe { NonZeroU8::new_unchecked(2) };
+pub const DEFAULT_NUMBER_OF_NAMES: NonZeroUsize = unsafe { NonZeroUsize::new_unchecked(6) };