+++ src/tests.rs
@@ -1,9 +1,11 @@
-use super::*;
 use axum::{
     body::Body,
     http::{Request, StatusCode},
 };
+use quickcheck_macros::quickcheck;
 use tower::{Service, ServiceExt};
+
+use super::*;
 
 #[tokio::test]
 async fn root() -> Result<()> {
@@ -19,13 +21,13 @@ async fn root() -> Result<()> {
 }
 
 #[tokio::test]
-async fn bad_request() -> Result<()> {
+async fn reject_negative() -> Result<()> {
     let mut app = app();
 
     let response = app
         .call(
             Request::builder()
-                .uri("/?words_per_name=-1")
+                .uri("/?words_per_name=-2")
                 .body(Body::empty())?,
         )
         .await?;
@@ -36,6 +38,33 @@ async fn bad_request() -> Result<()> {
         .call(
             Request::builder()
                 .uri("/?number_of_names=-1")
+                .body(Body::empty())?,
+        )
+        .await?;
+
+    assert_eq!(response.status(), StatusCode::BAD_REQUEST);
+
+    Ok(())
+}
+
+#[tokio::test]
+async fn reject_zero() -> Result<()> {
+    let mut app = app();
+
+    let response = app
+        .call(
+            Request::builder()
+                .uri("/?words_per_name=0")
+                .body(Body::empty())?,
+        )
+        .await?;
+
+    assert_eq!(response.status(), StatusCode::BAD_REQUEST);
+
+    let response = app
+        .call(
+            Request::builder()
+                .uri("/?number_of_names=0")
                 .body(Body::empty())?,
         )
         .await?;
@@ -105,4 +134,19 @@ async fn metrics() -> Result<()> {
     assert!(body.contains("http_requests_total{method=\"GET\",path=\"/metrics\",status=\"200\"}"));
 
     Ok(())
+}
+
+#[quickcheck]
+fn generated_names_are_unique(number_of_names: NonZeroU8) {
+    let number_of_names = NonZeroUsize::from(number_of_names);
+    assert_eq!(
+        generate_names(
+            HashSet::new(),
+            statics::DEFAULT_WORDS_PER_NAME,
+            statics::DEFAULT_SEPARATOR,
+            number_of_names
+        )
+        .len(),
+        number_of_names.get()
+    );
 }